refactor: condense account functionality
