+++ program/rust/src/tests/test_del_publisher.rs
@@ -1,15 +1,16 @@
 use {
     crate::{
-        c_oracle_header::{
+        accounts::{
             PriceAccount,
             PriceComponent,
             PriceInfo,
             PythAccount,
+        },
+        c_oracle_header::{
             PC_STATUS_TRADING,
             PC_VERSION,
         },
         deserialize::{
-            initialize_pyth_account_checked,
             load_checked,
             load_mut,
         },
@@ -56,7 +57,7 @@ fn test_del_publisher() {
 
     let mut price_setup = AccountSetup::new::<PriceAccount>(&program_id);
     let price_account = price_setup.as_account_info();
-    initialize_pyth_account_checked::<PriceAccount>(&price_account, PC_VERSION).unwrap();
+    PriceAccount::initialize(&price_account, PC_VERSION).unwrap();
     {
         let mut price_data = load_checked::<PriceAccount>(&price_account, PC_VERSION).unwrap();
         price_data.num_ = 1;
