+++ program/rust/src/tests/test_upd_aggregate.rs
@@ -1,14 +1,16 @@
 use {
     crate::{
-        c_oracle_header::{
+        accounts::{
             PriceAccount,
             PriceInfo,
+            PythAccount,
+        },
+        c_oracle_header::{
             PC_STATUS_TRADING,
             PC_STATUS_UNKNOWN,
             PC_VERSION,
         },
         deserialize::{
-            initialize_pyth_account_checked,
             load_checked,
             load_mut,
         },
@@ -65,7 +67,7 @@ fn test_upd_aggregate() {
     let mut price_setup = AccountSetup::new::<PriceAccount>(&program_id);
     let mut price_account = price_setup.as_account_info();
     price_account.is_signer = false;
-    initialize_pyth_account_checked::<PriceAccount>(&price_account, PC_VERSION).unwrap();
+    PriceAccount::initialize(&price_account, PC_VERSION).unwrap();
 
     // single publisher
     {
