+++ program/rust/src/tests/test_upd_price_no_fail_on_error.rs
@@ -1,13 +1,15 @@
 use {
     crate::{
-        c_oracle_header::{
+        accounts::{
             PriceAccount,
+            PythAccount,
+        },
+        c_oracle_header::{
             PC_STATUS_TRADING,
             PC_STATUS_UNKNOWN,
             PC_VERSION,
         },
         deserialize::{
-            initialize_pyth_account_checked,
             load_checked,
             load_mut,
         },
@@ -40,7 +42,7 @@ fn test_upd_price_no_fail_on_error_no_fail_on_error() {
     let mut price_setup = AccountSetup::new::<PriceAccount>(&program_id);
     let mut price_account = price_setup.as_account_info();
     price_account.is_signer = false;
-    initialize_pyth_account_checked::<PriceAccount>(&price_account, PC_VERSION).unwrap();
+    PriceAccount::initialize(&price_account, PC_VERSION).unwrap();
 
     let mut clock_setup = AccountSetup::new_clock();
     let mut clock_account = clock_setup.as_account_info();
