+++ program/rust/src/time_machine_types.rs
@@ -2,9 +2,11 @@
 #![allow(dead_code)]
 use {
     crate::{
-        c_oracle_header::{
+        accounts::{
             PriceAccount,
             PythAccount,
+        },
+        c_oracle_header::{
             EXTRA_PUBLISHER_SPACE,
             PC_ACCTYPE_PRICE,
             PC_MAX_SEND_LATENCY,
